requirements.txt
torch==2.3.1
transformers==4.41.0
peft==0.10.0
accelerate==0.30.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
fastapi==0.111.0
uvicorn==0.30.0
streamlit==1.36.0
numpy==1.26.4
scipy==1.13.1
requests==2.32.3
pytest==8.2.2
slowapi==0.1.9
anyio==4.4.0
prometheus-client==0.20.0
ruff==0.5.7
mypy==1.11.1
types-requests==2.32.0.20240712
locust==2.29.1
pydantic-settings>=2.3.0
ragas>=0.1.14
mlflow>=2.14.0
datasets>=2.20.0
rank-bm25>=0.2.2